Travis CI

+++ install.sh
@@ -21,10 +21,13 @@ COUCHDB_DB="foodb"
 ## Misc settings
 
 ES_VERSION=0.20.5
+MINIMAL_INSTALL=
+JDK=1
 
 if [ ! -f jdk.tar.gz ]; then
-    echo "Please read the top of this file."
-    exit 1
+    echo "WARNING: No JDK tarball found; some pieces of CommCareHQ (Cloudcare) may be nonfunctional"
+    JDK=
+    MINIMAL_INSTALL=1
 fi
 
 ## Install OS-level package dependencies
@@ -106,7 +109,7 @@ if [[ ! $(grep virtualenvwrapper ~/.bashrc) ]]; then
 fi
 
 ## Install Java ##
-if [ ! -d /usr/lib/jvm/jdk1.7.0 ]; then
+if [ "$JDK" ] && [ ! -d /usr/lib/jvm/jdk1.7.0 ]; then
     tar -xzf jdk.tar.gz
     sudo mkdir /usr/lib/jvm
     sudo rm -r /usr/lib/jvm/jdk1.7.0/
@@ -121,7 +124,7 @@ if [ ! -d /usr/lib/jvm/jdk1.7.0 ]; then
 fi
 
 ## Install Jython ##
-if [ ! -d /usr/local/lib/jython ]; then
+if [ "$JDK" ] && [ ! -d /usr/local/lib/jython ]; then
     if [ ! -f jython_installer-2.5.2.jar ]; then
         wget http://downloads.sourceforge.net/project/jython/jython/2.5.2/jython_installer-2.5.2.jar
     fi
@@ -175,7 +178,7 @@ if [ ! -f /etc/init.d/couchdb ]; then
 fi
 
 ## Install couchdb-lucene
-if [ ! -f /etc/init.d/couchdb-lucene ]; then
+if [ ! "$MINIMAL_INSTALL" ] && ! -f /etc/init.d/couchdb-lucene ]; then
     if [ ! -f v0.8.0.zip ]; then
         wget https://github.com/rnewson/couchdb-lucene/archive/v0.8.0.zip
     fi
@@ -185,7 +188,7 @@ if [ ! -f /etc/init.d/couchdb-lucene ]; then
     sudo cp /usr/local/couchdb-lucene-0.8.0/tools/etc/init.d/couchdb-lucene/couchdb-lucene /etc/init.d/
 fi
 
-if [[ ! $(grep _fti /usr/local/etc/couchdb/local.ini) ]]; then
+if [ -e /usr/local/etc/couchdb/local.ini ] && [[ ! $(grep _fti /usr/local/etc/couchdb/local.ini) ]]; then
     config=/usr/local/etc/couchdb/local.ini
     sudo sed -i '/\[couchdb\]/ a\os_process_timeout=60000' $config
 
@@ -199,7 +202,7 @@ _fti = {couch_httpd_external, handle_external_req, <<\"fti\">>}
 fi
 
 ## Install elastic-search ##
-if [ ! -f /etc/init.d/elasticsearch ]; then
+if [ "$JDK" ] && [ ! -f /etc/init.d/elasticsearch ]; then
     if [ "$PM" = "apt-ubuntu" ]; then
         file=elasticsearch-$ES_VERSION.deb
         if [ ! -f $file ]; then
@@ -236,40 +239,44 @@ fi
 # installs a system java package and changes the configured java install path,
 # which we don't want
 
-sudo update-alternatives --install "/usr/bin/java" "java" "/usr/lib/jvm/jdk1.7.0/bin/java" 1
-sudo update-alternatives --install "/usr/bin/javac" "javac" "/usr/lib/jvm/jdk1.7.0/bin/javac" 1
-sudo update-alternatives --install "/usr/bin/javaws" "javaws" "/usr/lib/jvm/jdk1.7.0/bin/javaws" 1
+if [ ! "$MINIMAL_INSTALL" ]; then
+    sudo update-alternatives --install "/usr/bin/java" "java" "/usr/lib/jvm/jdk1.7.0/bin/java" 1
+    sudo update-alternatives --install "/usr/bin/javac" "javac" "/usr/lib/jvm/jdk1.7.0/bin/javac" 1
+    sudo update-alternatives --install "/usr/bin/javaws" "javaws" "/usr/lib/jvm/jdk1.7.0/bin/javaws" 1
+fi
 
 sudo update-alternatives --config java
 
 ## Ensure services start on startup ##
-if [ "$PM" = "apt-ubuntu" ]; then
-    sudo update-rc.d couchdb defaults
-    sudo update-rc.d couchdb-lucene defaults
+if [ ! "$MINIMAL_INSTALL" ]; then
+    if [ "$PM" = "apt-ubuntu" ]; then
+        sudo update-rc.d couchdb defaults
+        sudo update-rc.d couchdb-lucene defaults
 
-    # these should already be on by default
-    sudo update-rc.d elasticsearch defaults
-    sudo update-rc.d memcached defaults
-    sudo update-rc.d postgresql defaults
-elif [ "$PM" = "yum-rhel" ]; then
-    sudo chkconfig --add couchdb
-    sudo chkconfig --add elasticsearch
-    sudo chkconfig --add memcached
-    sudo chkconfig --add postgresql
-    sudo chkconfig --add couchdb-lucene
+        # these should already be on by default
+        sudo update-rc.d elasticsearch defaults
+        sudo update-rc.d memcached defaults
+        sudo update-rc.d postgresql defaults
+    elif [ "$PM" = "yum-rhel" ]; then
+        sudo chkconfig --add couchdb
+        sudo chkconfig --add elasticsearch
+        sudo chkconfig --add memcached
+        sudo chkconfig --add postgresql
+        sudo chkconfig --add couchdb-lucene
+    
+        sudo chkconfig couchdb on
+        sudo chkconfig elasticsearch on
+        sudo chkconfig memcached on
+        sudo chkconfig postgresql on
+        sudo chkconfig couchdb-lucene on
+    fi
 
-    sudo chkconfig couchdb on
-    sudo chkconfig elasticsearch on
-    sudo chkconfig memcached on
-    sudo chkconfig postgresql on
-    sudo chkconfig couchdb-lucene on
+    ## Ensure services are running ##
+    sudo service couchdb start
+    sudo service elasticsearch start
+    sudo service memcached start
+    sudo service postgresql start
 fi
-
-## Ensure services are running ##
-sudo service couchdb start
-sudo service elasticsearch start
-sudo service memcached start
-sudo service postgresql start
 
 ## Configure databases ##
 DB=$POSTGRES_DB
